+++ src/TSID/include/BipedalLocomotion/TSID/JointDynamicsTask.h
@@ -46,7 +46,8 @@ class JointDynamicsTask : public TSIDLinearTask
     System::VariablesHandler::VariableDescription m_jointsTorqueVariable; /**< Variable
                                                                              describing the joint
                                                                              torques */
-
+    bool m_useMassMatrixRegularizationTerm{false}; /**< True if the mass matrix regularization is used*/
+    Eigen::MatrixXd m_massMatrixRegularizationTerm; /**< Variable storing mass matrix regularization term*/
     struct ContactWrench
     {
         iDynTree::FrameIndex frameIndex; /**< Frame used to express the contact wrench */
@@ -107,6 +108,16 @@ public:
      * @return True in case of success, false otherwise.
      */
     bool setVariablesHandler(const System::VariablesHandler& variablesHandler) override;
+
+    /**
+     * Set the mass matrix regularization term. i.e. \f$\bar{M} = M + M _ {reg}\f$. Where  \f$M\f$
+     * is the mass matrix and  \f$M_{reg}\f$ is the matrix regularization term.
+     * @param matrix the regularization term for the mass matrix.
+     * @notice Calling this function is not mandatory. Call it only if you want to add a
+     * regularization term.
+     * @return true in case of success, false otherwise.
+     */
+    bool setMassMatrixRegularization(const Eigen::Ref<const Eigen::MatrixXd>& matrix);
 
     /**
      * Update the content of the element.